pyproject.toml
[tool.poetry]
name = "hw30-django-users-and-roles"
version = "0.1.0"
description = ""
authors = ["samsegomof <wemagonof91@gmail.com>"]
readme = "README.md"
packages = [{include = "hw30_django_users_and_roles"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
psycopg2 = "^2.9.5"
djangorestframework = "^3.14.0"
pillow = "^9.3.0"
djangorestframework-simplejwt = "^5.2.2"
python-dateutil = "^2.8.2"
pytest-factoryboy = "^2.5.0"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
